refactored such that notifier and data_platform are now top-level modules
