+++ src/__init__.py
@@ -1,0 +1,1 @@
+# Make data_platforms a top-level package
