+++ src/data_platform/__init__.py
@@ -1,0 +1,3 @@
+# flake8: noqa
+from src.data_platform.demo import *
+from src.data_platform.dummy import *
