+++ src/data_platform/demo/__init__.py
@@ -1,0 +1,12 @@
+# flake8: noqa
+from src.data_platform.demo.demo_query_handler import (
+    DemoQueryHandler,
+    DefaultDemoTranslator
+)
+from src.data_platform.demo.demo_naming_resolver import (
+    DemoNamingResolver
+)
+from src.data_platform.demo.demo_platform import DemoDataPlatform
+from src.data_platform.demo.demo_platform_factory import (
+    DemoDataPlatformFactory
+)
