+++ src/data_platform/dummy/__init__.py
@@ -1,0 +1,3 @@
+# flake8: noqa
+from src.data_platform.dummy.dummy_platform import DummyPlatform
+from src.data_platform.dummy.dummy_platform_factory import DummyPlatformFactory
