--- src/domain_config/di/__init__.py
@@ -1,2 +1,0 @@
-# flake8: noqa
-from src.domain_config.di.di import DependencyInjector
