+++ src/notifier/__init__.py
@@ -1,0 +1,3 @@
+# flake8: noqa
+from src.notifier.in_memory_notifier import InMemoryNotifier
+from src.notifier.stdout_notifier import StdoutNotifier
