+++ src/testcase/adapters/data_platform.py
@@ -1,0 +1,4 @@
+"""
+Implements the IDataPlatform interface is moved to be a top-level module since
+it is used by other hexagons (i.e. top level modules)
+"""
