--- src/testcase/adapters/data_platforms/__init__.py
@@ -1,3 +1,0 @@
-# flake8: noqa
-from src.testcase.adapters.data_platforms.demo import *
-from src.testcase.adapters.data_platforms.dummy import *
