--- src/testcase/adapters/data_platforms/demo/__init__.py
@@ -1,12 +1,0 @@
-# flake8: noqa
-from src.testcase.adapters.data_platforms.demo.demo_query_handler import (
-    DemoQueryHandler,
-    DefaultDemoTranslator
-)
-from src.testcase.adapters.data_platforms.demo.demo_naming_resolver import (
-    DemoNamingResolver
-)
-from src.testcase.adapters.data_platforms.demo.demo_platform import DemoDataPlatform
-from src.testcase.adapters.data_platforms.demo.demo_platform_factory import (
-    DemoDataPlatformFactory
-)
