--- src/testcase/adapters/data_platforms/dummy/__init__.py
@@ -1,3 +1,0 @@
-# flake8: noqa
-from src.testcase.adapters.data_platforms.dummy.dummy_platform import DummyPlatform
-from src.testcase.adapters.data_platforms.dummy.dummy_platform_factory import DummyPlatformFactory
