+++ src/testcase/adapters/notifier.py
@@ -1,0 +1,4 @@
+"""
+Implementation of the INotifier interface is moved to be a top-level module since
+it is used by other hexagons (i.e. top level modules)
+"""
