--- src/testcase/adapters/notifiers/__init__.py
@@ -1,3 +1,0 @@
-# flake8: noqa
-from src.testcase.adapters.notifiers.in_memory_notifier import InMemoryNotifier
-from src.testcase.adapters.notifiers.stdout_notifier import StdoutNotifier
