+++ src/testcase/ports/__init__.py
@@ -1,6 +1,15 @@
 # flake8: noqa
-from src.testcase.ports.i_data_platform import IDataPlatform
-from src.testcase.ports.i_data_platform_factory import IDataPlatformFactory
-from src.testcase.ports.i_notifier import INotifier
-from src.testcase.ports.i_run_testcases import (
+from src.testcase.ports.infrastructure.data_platform.i_data_platform import IDataPlatform
+from src.testcase.ports.infrastructure.data_platform.i_data_platform_factory import IDataPlatformFactory
+from src.testcase.ports.infrastructure.notifier.i_notifier import INotifier
+from src.testcase.ports.drivers.i_run_testcases import (
     IRunTestCasesCommandHandler, RunTestCaseCommand)
+
+# Re-export for easier imports
+__all__ = [
+    "IDataPlatform",
+    "IDataPlatformFactory",
+    "INotifier",
+    "IRunTestCasesCommandHandler",
+    "RunTestCaseCommand",
+]
