+++ src/testcase/ports/infrastructure/data_platform/i_data_platform.py
